Badoum ! Euh entre autre background des icones de catégories et Fab transparent

+++ app/src/main/java/com/timappweb/timapp/activities/DrawerActivity.java
@@ -334,7 +334,7 @@ public class DrawerActivity extends BaseActivity implements NavigationView.OnNav
         }
     }
 
-    public void setFabOpacity(int opacity) {
+    public void setFabOpacity(float opacity) {
         addSpotFloatingButton.setAlpha(opacity);
     }
